pyproject.toml
[tool.poetry]
name = "clusterexp"
version = "0.1.0"
description = "Clustering Experiments"
authors = ["Natacha, Galmiche <natacha.galmiche@uib.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pycvi-lib = {git = "https://github.com/nglm/pycvi.git",  branch = "master"}
scikit-learn-extra = "^0.3.0"
matplotlib = "^3.8.4"

[tool.poetry.group.notebook]
optional = true

[tool.poetry.group.notebook.dependencies]
ipykernel = "^6.27.1"
pandas = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
